Add Use More OOP
