--- data/duke.txt
@@ -1,2 +1,0 @@
-E | true | hello | 2022-12-20 1800
-D | false | deadline  | 2022-12-20 1800
