+++ src/main/java/DeadlineTask.java
@@ -41,7 +41,7 @@ public class DeadlineTask extends Task {
     @Override
     public String toString() {
         DateTimeFormatter formatterDate = DateTimeFormatter.ofPattern("MMM dd yyyy");
-        DateTimeFormatter formatterTime =DateTimeFormatter.ofPattern("h:mma");
+        DateTimeFormatter formatterTime =DateTimeFormatter.ofPattern("h:mm a");
         return "[D]" + super.toString() + " (by: "
                 + this.date.format(formatterDate) + " "
                 + this.time.format(formatterTime) + ")";
